+++ CHANGELOG.md
@@ -41,6 +41,8 @@ SPDX-License-Identifier: CC0-1.0
 - Add the `PageDecorator` trait, the `SimplePageDecorator` implementation and
   the `Document::set_page_decorator` method to allow customization of all
   document pages.
+- Add support for kerning and add the `Font::kerning` and `Font::glyph_ids`
+  methods.
 
 # v0.1.1 (2020-10-16)